# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Together team collaboration platform API for managing AI coding tools (Claude Code, Codex, OpenCode). Provides centralized provider management, team-level settings, usage statistics, and automatic configuration distribution.",
        "title": "Code Together API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/code-together/ai-together/issues",
            "email": "support@ai-together.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/setup/admin": {
            "post": {
                "description": "Create the initial admin account, tenant, and team (only works when no users exist)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup"
                ],
                "summary": "Create initial admin",
                "parameters": [
                    {
                        "description": "Admin setup data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SetupAdminRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SetupAdminResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Setup already completed",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/setup/status": {
            "get": {
                "description": "Check if initial setup is required (no users exist)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setup"
                ],
                "summary": "Get setup status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SetupStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.SetupAdminRequest": {
            "type": "object",
            "required": [
                "admin_email",
                "admin_name",
                "admin_password",
                "organization_name"
            ],
            "properties": {
                "admin_email": {
                    "type": "string"
                },
                "admin_name": {
                    "type": "string"
                },
                "admin_password": {
                    "type": "string",
                    "minLength": 8
                },
                "organization_name": {
                    "type": "string"
                }
            }
        },
        "handlers.SetupAdminResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "team": {
                    "type": "object",
                    "additionalProperties": true
                },
                "tenant": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "handlers.SetupStatusResponse": {
            "type": "object",
            "properties": {
                "setup_required": {
                    "type": "boolean"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "details": {
                    "description": "Additional context (debug mode only)",
                    "type": "string"
                },
                "error": {
                    "description": "User-friendly error message",
                    "type": "string"
                },
                "request": {
                    "description": "Request ID for support tracking",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}